% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{calculates mean of sparse numeric vector object}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric, the mean of the sparse numeric
}
\description{
calculates mean of sparse numeric vector object
}
