% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{gets squared norm of vector (sqrt of sum of squared elements)}
\usage{
\S4method{norm}{sparse_numeric}(object, ...)
}
\arguments{
\item{object}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric, the norm of the sparse numeric
}
\description{
gets squared norm of vector (sqrt of sum of squared elements)
}
