% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{norm}
\alias{norm}
\title{gets squared norm of vector (sqrt of sum of squared elements)}
\usage{
norm(object, ...)
}
\arguments{
\item{object}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric, the norm of the sparse numeric
}
\description{
gets squared norm of vector (sqrt of sum of squared elements)
}
