% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{creates scatterplot of overlapping nonzero positions between two sparse numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{the first sparse numeric vector object}

\item{y}{the second sparse numeric vector object}

\item{...}{additional arguments}
}
\description{
creates scatterplot of overlapping nonzero positions between two sparse numeric vectors
}
