% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{+,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{adds together two sparse numeric vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{the first sparse numeric vector object}

\item{e2}{the second sparse numeric vector object}
}
\value{
a sparse numeric vector object, the sum of the two passed in
}
\description{
adds together two sparse numeric vectors
}
