% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{prints out attributes of sparse numeric vector object
(positions of nonzero elements, values of nonzero elements, length of vector)}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{the sparse numeric vector object}
}
\description{
prints out attributes of sparse numeric vector object
(positions of nonzero elements, values of nonzero elements, length of vector)
}
