% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{sparse_add}
\alias{sparse_add}
\title{adds together two sparse numeric vectors}
\usage{
sparse_add(object, object2, ...)
}
\arguments{
\item{object}{the first sparse numeric vector object}

\item{object2}{the second sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a sparse numeric vector object, the sum of the two passed in
}
\description{
adds together two sparse numeric vectors
}
