% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{gets the cross product of two sparse numeric vectors}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(object, object2, ...)
}
\arguments{
\item{object}{the first sparse numeric vector object}

\item{object2}{the second sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric, the cross product of the two vectors
}
\description{
gets the cross product of two sparse numeric vectors
}
