% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{gets the cross product of two sparse numeric vectors}
\usage{
sparse_crossprod(object, object2, ...)
}
\arguments{
\item{object}{the first sparse numeric vector object}

\item{object2}{the second sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric, the cross product of the two vectors
}
\description{
gets the cross product of two sparse numeric vectors
}
