% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{multiplies together two sparse numeric vectors}
\usage{
sparse_mult(object, object2, ...)
}
\arguments{
\item{object}{the first sparse numeric vector object}

\item{object2}{the second sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a sparse numeric vector object, the product of the two passed in
}
\description{
multiplies together two sparse numeric vectors
}
