% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{sparse numeric vector class}
\description{
represents sparse numeric vectors by storing only nonzero positions and values
}
\details{
value and pos vectors must be of the same length, they correspond to each other
}
\section{Slots}{

\describe{
\item{\code{value}}{numeric vector of values for nonzero elements in sparse numeric}

\item{\code{pos}}{integer vector of positions for nonzero elements in sparse numeric}

\item{\code{length}}{total integer length of sparse numeric vector, including all zeros}
}}

