% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{-,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{subtracts one sparse numeric vector from another}
\usage{
\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{the sparse numeric vector object to be subtracted from}

\item{e2}{the sparse numeric vector object to subtract}
}
\value{
a sparse numeric vector object, result of object - object2
}
\description{
subtracts one sparse numeric vector from another
}
