% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{subtracts one sparse numeric vector from another}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(object, object2, ...)
}
\arguments{
\item{object}{the sparse numeric vector object to be subtracted from}

\item{object2}{the sparse numeric vector object to subtract}

\item{...}{additional arguments}
}
\value{
a sparse numeric vector object, result of object - object2
}
\description{
subtracts one sparse numeric vector from another
}
