% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{standardize}
\alias{standardize}
\title{gets standardized version of sparse numeric vector object, in dense form
takes each element of vector and subtracts vector mean & divides by vector st. dev}
\usage{
standardize(object, ...)
}
\arguments{
\item{object}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a dense vector, the standardized version of the sparse numeric
}
\description{
gets standardized version of sparse numeric vector object, in dense form
takes each element of vector and subtracts vector mean & divides by vector st. dev
}
