% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{unsparse,sparse_numeric-method}
\alias{unsparse,sparse_numeric-method}
\title{creates numeric vector with all zeros removed from sparse numeric object}
\usage{
\S4method{unsparse}{sparse_numeric}(object, ...)
}
\arguments{
\item{object}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric vector, no longer sparse and no 0s
}
\description{
creates numeric vector with all zeros removed from sparse numeric object
}
