% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenumeric.R
\name{unsparse}
\alias{unsparse}
\title{creates numeric vector with all zeros removed from sparse numeric object}
\usage{
unsparse(object, ...)
}
\arguments{
\item{object}{the sparse numeric vector object}

\item{...}{additional arguments}
}
\value{
a numeric vector
}
\description{
creates numeric vector with all zeros removed from sparse numeric object
}
